remove dtype

+++ QCLayers.py
@@ -145,7 +145,7 @@ class QCLayers(object):
         # create two lists to store start ind and end ind
 
         self.eigenEs = np.empty((0))
-        self.psis = np.empty((0, self.xPoints.size), dtype=np.float64)
+        self.psis = np.empty((0, self.xPoints.size))
         for n in range(0, len(StartInd)):
             dCL = copy.deepcopy(self)
             dCL.repeats = 1
@@ -167,7 +167,7 @@ class QCLayers(object):
             xInd = (self.xLayerNums >= StartInd[n]) & \
                    (self.xLayerNums < EndInd[n])
 
-            psis = np.zeros((dCL.eigenEs.shape[0], 0), dtype=np.float64)
+            psis = np.zeros((dCL.eigenEs.shape[0], 0))
             for j in range(0, self.repeats):
                 # this is where to add global Efield
                 psis = np.concatenate((psis, dCL.psis), axis=1)
